Strap yourselves in kids, because I am about to spit fire. Our boolean handling code was total shit. This commit fixes a lot of the problems:
(1) We were not able to convert VARCHARs to BOOLEANs.
(2) And then even if we could, we had mispelled 'true' and 'false' so it wouldn't have worked anyway.
(3) BooleanType::CompareEquals would always return true if you compared a NULL value...
(4) This was because it would call ValueFactory::GetBooleanValue() which only supported bool inputs. That means PELOTON_BOOLEAN_NULL was always converted to true
(5) And then because I was disappointed that our code coverage has not moved up in a few months, I busted out new boolean test cases to make sure that all of this is working properly.

I did this in just three hours today in between meetings. Where is your god now?