Fuck yes. Here we go now.
This is our new TempTable object. It's meant to be more lightweight than DataTable. It's not thread-safe. It's not durable. But that's the way it's supposed to be. We have a test case that inserts some tuples and then checks to see whether we get correct value back.

The next step is to then go change AggregateExecutor to use this instead of DataTable. #183