Explicitly make `pelaton` depend on custom target, `libpg_query`
This doesn't work on my system most of the times. I think this is
because of a race condition involved here. add_custom_target makes sure
that target will be built, but doesn't specify when. By adding explicit
dependency, we make sure that by the time we link pelaton with
libpg_query, libpg_query is built.

With this patch, it always work for me.