Fix superfluous and incorrect capnp cmake exports
Some of these variables are already exported by cmake files in capnproto
in third_party. And correctly.

The problem I see in my machine is that capnproto writes the static libs
to prefix/lib64/ but hardcoded CAPNP_LIBRARIES says that they are in the
prefix/lib/ directory. This makes peloton fail to link with capnp static
libs.

Let capnp cmake files in third_party handle it correctly.